\name{rareGE-package}
\alias{rareGE-package}
\docType{package}
\title{
Testing Gene-Environment Interaction for Rare Genetic Variants
}
\description{
Tests gene-environment interaction for rare genetic variants using Sequence Kernel Association Test (SKAT) type gene-based tests. Includes two tests for the interaction term only, and one joint test for genetic main effects and gene-environment interaction.
}
\details{
\tabular{ll}{
Package: \tab rareGE\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-07-28\cr
License: \tab GPL-3\cr
}
}
\author{
Han Chen

Maintainer: Han Chen <hanchen@hsph.harvard.edu>
}
\references{
Chen H, Meigs JB, Dupuis J. (2014) Incorporating gene-environment interaction in testing for association with rare genetic variants. Hum Hered 78, 81-90.
}
\keyword{ package }
\seealso{
\code{\link{INT_FIX}}, \code{\link{INT_RAN}}, \code{\link{JOINT}}, \code{\link{rareGE}}
}
