\name{rareGEgeno}
\alias{rareGEgeno}
\docType{data}
\title{
Example genotype data
}
\description{
Example genotype data for rareGE.
}
\usage{data(rareGEgeno)}
\format{
A data matrix for 500 individuals and a gene with 20 SNPs. Each row represents an individual and each column represents a SNP.
}
\keyword{datasets}
