\name{rareGEpheno}
\alias{rareGEpheno}
\docType{data}
\title{
Example phenotype data
}
\description{
Example phenotype data for rareGE.
}
\usage{data(rareGEpheno)}
\format{
  A data frame with 500 observations on the following 5 variables.
  \describe{
    \item{\code{y1}}{a quantitative trait.}
    \item{\code{y2}}{a dichotomous trait.}
    \item{\code{age}}{a quantitative covariate.}
    \item{\code{sex}}{a dichotomous covariate.}
    \item{\code{bmi}}{a quantitative covariate.}
  }
}
\keyword{datasets}
